100-atoi.c

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -23,5 +23,5 @@ break;
 i++;
 }
 
-return (result *sign);
+return ((int)(result *sign));
 }
